app/backend/pyproject.toml
[project]
name = "local-ai-toolkit-backend"
version = "0.1.0"
description = "FastAPI backend scaffold for the local AI toolkit"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "pydantic-settings>=2.10.0",
  "python-multipart>=0.0.20",
  "uvicorn>=0.35.0",
]

[tool.uv]
package = false
